handle deprecated/versionadded tags
Translate @since to Sphinx's '.. versionadded::' directive. Likewise,
translate @deprecated to '.. deprecated::' directive. The former is
trivial and can be translated 1:1. The latter requires some trickery
as the Spinx deprecated directive requires a mandatory version
argument which somehow has to be extracted from @deprecated.

For now, we just look at he first word of the @deprecated text and
check if it matches the format of a "version number". If it does, we
use it. If it doesn't, fallback to a hand-written deprecated
statement.